fix: make sure transactions add a memo instruction if enabled (#41)
