+++ src/kinetic_sdk/helpers/generate_make_transfer_batch_transaction.py
@@ -1,52 +1,22 @@
 from typing import List, Dict
+
 from solana.publickey import PublicKey
-
-from solders.instruction import AccountMeta, Instruction
+from solders.instruction import Instruction
 from solders.message import Message as SoldersMessage
-from solders.pubkey import Pubkey
-
-from spl.token._layouts import INSTRUCTIONS_LAYOUT, InstructionType
 from spl.token.instructions import get_associated_token_address
 
+from kinetic_sdk.helpers.create_make_transfer_instruction import create_make_transfer_instruction
+from kinetic_sdk.helpers.create_memo_instruction import create_memo_instruction
 from kinetic_sdk.helpers.sign_and_serialize_transaction import sign_and_serialize_transaction
 from kinetic_sdk.models.public_key_string import PublicKeyString
 from kinetic_sdk.models.transaction_type import TransactionType
 
-from kinetic_sdk.models.constants import TOKEN_PROGRAM_ID
-
-
-def create_make_transfer_instruction(
-    source: Pubkey,
-    source_token_account: Pubkey,
-    destination_token_account: Pubkey,
-    mint: Pubkey,
-    amount: int,
-    decimals: int,
-):
-    account_metas = [
-        AccountMeta(source_token_account, False, True),
-        AccountMeta(mint, False, False),
-        AccountMeta(destination_token_account, False, True),
-        AccountMeta(source, True, False),
-    ]
-
-    amount = amount * 10 ** decimals
-    data = INSTRUCTIONS_LAYOUT.build(
-        dict(instruction_type=InstructionType.TRANSFER2, args=dict(amount=amount, decimals=decimals))
-    )
-
-    return Instruction(
-        program_id=TOKEN_PROGRAM_ID,
-        data=data,
-        accounts=account_metas
-    )
-
 
 def generate_make_transfer_batch_transaction(
     add_memo: bool,
-    app_index: int,
     destinations: List[Dict[PublicKeyString, str]],
     decimals: int,
+    index: int,
     mint_fee_payer: str,
     mint_public_key: str,
     recent_blockhash: str,
@@ -54,6 +24,9 @@ def generate_make_transfer_batch_transaction(
     tx_type: TransactionType = TransactionType.NONE
 ):
     instructions: List[Instruction] = []
+
+    if add_memo:
+        instructions.append(create_memo_instruction(index, tx_type))
 
     for destination in destinations:
         source_token_account = get_associated_token_address(source.public_key, PublicKey(mint_public_key))
